Add iterator of datastories for the selected municipality, for navigation.
Also rework to use a django paginator object (they do work with one item
per page).  The original behavior is emulated, so the template hasn't
changed, but there are paginator and paginator_page context variables,
as well as the datastories variable for the iterator of the previous
paragraph, in addition to the original variables.  If the template gets
reworked to use the paginator, the old page, page_count, page_next,
page_prev, and page_num could be eliminated.